refactor: Cars 클래스 개선
