--- src/main/java/domain/Cars.java
@@ -1,24 +1,0 @@
-package domain;
-
-import java.util.Collections;
-import java.util.List;
-
-public class Cars {
-    private final List<Car> cars;
-
-    private Cars(List<Car> cars) {
-        this.cars = cars;
-    }
-
-    public static Cars from(List<Car> cars) {
-        return new Cars(cars);
-    }
-
-    public List<Car> getCars() {
-        return Collections.unmodifiableList(cars);
-    }
-
-    public void attemptToMove() {
-        cars.forEach(Car::attemptToMove);
-    }
-}
